### 2018-09-25 #976 @RichardLitt
I don't think I would have the requisite experience to judge this adequately from a digital humanities perspective, at the moment.
