### 2018-10-21 #976 @yochannah
Still trying to chase down some reviewers for this. Thanks for your patience @davekinkead!

@BasTesterink - I note on the JOSS reviewer list that one of your domain areas is formal logic. Would you be interested in reviewing this package?
