### 2018-10-21 #976 @davekinkead
Thanks for keeping us in the loop.

> On 22 Oct 2018, at 8:17 am, Yo Yehudi <notifications@github.com> wrote:
>
> Still trying to chase down some reviewers for this. Thanks for your patience @davekinkead <https://github.com/davekinkead>!
>
> @BasTesterink <https://github.com/BasTesterink> - I note on the JOSS reviewer list that one of your domain areas is formal logic. Would you be interested in reviewing this package?
>
