### 2018-10-22 #976 @mbod
I'm happy to take a look at this - I don't have formal logic expertise but
have linguistics, digital humanities and JS knowledge

On Sun, Oct 21, 2018 at 6:20 PM Dave Kinkead <notifications@github.com>
wrote:

> Thanks for keeping us in the loop.
>
> > On 22 Oct 2018, at 8:17 am, Yo Yehudi <notifications@github.com> wrote:
> >
> > Still trying to chase down some reviewers for this. Thanks for your
> patience @davekinkead <https://github.com/davekinkead>!
> >
> > @BasTesterink <https://github.com/BasTesterink> - I note on the JOSS
> reviewer list that one of your domain areas is formal logic. Would you be
> interested in reviewing this package?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub <
> https://github.com/openjournals/joss-reviews/issues/976#issuecomment-431709087>,
> or mute the thread <
> https://github.com/notifications/unsubscribe-auth/AAlWJUmcj7Y08lFsQ-qx-H68LOqFpRHsks5unPJ1gaJpZM4W3JvR
> >.
> >
>
