### 2018-10-22 #1044 @whedon
[REVIEW]: Reasons: A digital argument mapping library for modern browsers
**Submitting author:** @davekinkead (<a href="http://orcid.org/0000-0001-5396-8099">David Kinkead</a>)
**Repository:** <a href="https://github.com/davekinkead/reasons" target ="_blank">https://github.com/davekinkead/reasons</a>
**Version:** v1.0.1
