### 2018-10-23 davekinkead/reasons#1 @andytwoods
review issue. Update installation instructions.
As per https://github.com/openjournals/joss-reviews/issues/1044 Functionality, Installation.

I couldnt get this to work:
https://github.com/davekinkead/reasons#embedding-reasons

I got it to work this the below. Perhaps you want to add an eg div instead of body though.
```
<html>
<body>
  <script src="reasons.js"></script>
  <script>
    let graph = [
        {id: 'p1', text: "Circular arguments work"},
        {id: 'c1', text: "Circular arguments work"},
        {from: 'p1', to: 'c1', type: "because"},
        {from: 'c1', to: 'p1', type: "because"}
      ]
    Reasons.mapper('body').render(graph)
  </script>
</body>
</html>
```
