### 2018-10-23 davekinkead/reasons#2 @andytwoods
Review issue: functionality
As per https://github.com/openjournals/joss-reviews/issues/1044 Functionality, functionality.

on windows, in chrome, I couldn't undo via control-z.
