### 2018-10-23 davekinkead/reasons#3 @andytwoods
review issue: Automated tests
As per https://github.com/openjournals/joss-reviews/issues/1044

Can you explicitly spell out how to run the tests. Ideally I could copy/paste.
