### 2018-10-23 davekinkead/reasons#4 @andytwoods
review issue: community guidelines
As per https://github.com/openjournals/joss-reviews/issues/1044

Please add text as suggested:
"Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"
