### 2018-10-23 b10e1ba
Update readme 'Embedding Reasons'. Closes #1
