### 2018-10-23 54844fa
Update readme 'Development & Testing'
