### 2018-10-23 davekinkead/reasons#1 @davekinkead
Thanks @andytwoods for taking the time to review this software.

The `Reasons.mapper('element')` function accepts any valid DOM reference.  That could be a `<body>` tag with `Reasons.mapper('body')` or a `<div id="map">` tag with `Reasons.mapper('#map')`.

The readme has been updated to make this more explicit.
