### 2018-10-23 davekinkead/reasons#3 @davekinkead
Test instructions have been updated in 54844fa8.

Try the following three commands from the project dir...

`npm install`
`npm install --only=dev`
`npm test`

The first installs production dependencies required to build the software.  The second the dependencies to run the tests. The third to trigger the automated test suite.

None of this is required to actually use the software however.
