### 2018-10-23 3c063b7
Update readme 'Contributing & Support'. Closes #4
