### 2018-10-23 davekinkead/reasons#2 @davekinkead
Thanks for the bug report @andytwoods

I can't reproduce the problem on my setup but the relevant code is here https://github.com/davekinkead/reasons/blob/3c063b76f9f6c6e28137833cd43a38e10bf1b9f9/lib/ui.js#L138-L140

`Cmd + z`, `Ctrl + z`, or the `browser specific metakey + z` should all work.  I'll try to find a windows machine with chrome shortly.

In the mean time, can you please try the following http://www.reasons.io/maps/e25636f to see if the issue is browser or environment caused.
