### 2018-10-23 davekinkead/reasons#2 @davekinkead
@andytwoods could you also please visit https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/metaKey and see if you can get `metaKey = true` in the [example](https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/metaKey#Example)
