### 2018-10-24 davekinkead/reasons#2 @andytwoods
I'm a reviewer for your submission to joss. I don't use your software. Not sure it's fair I am also involved in the debug process. Just use VirtualBox to check out the issue?
