### 2018-11-05 davekinkead/reasons#2 @davekinkead
Quick update - I'm currently at a few conferences and will address this later this month when I'm back
