### 2018-12-14 #1044 @davekinkead
Thanks Yo - I’ve just got back from conference season and am working on an issue raised by the reviewer.

Dave

> On 14 Dec 2018, at 9:32 am, Yo Yehudi <notifications@github.com> wrote:
>
> Quick status check - is there anything I can do to help progress this? 👍
>
