### 2019-01-13 #1044 @yochannah
Quick status check - let me know when we're ready to move forward or if you have any issues :)
