### 2019-02-14 #1044 @davekinkead
Hi all - I’m overseas at the moment but will be back on campus next week to action anything outstanding.

Dave

Sent from my iPhone

> On 13 Feb 2019, at 10:49 pm, Yo Yehudi <notifications@github.com> wrote:
>
> thanks @labarba! I think @davekinkead is also working on the issues raised by @andytwoods and will let us know when he is ready to revisit.
>
