+++ lib/view.js
@@ -1,5 +1,5 @@
 //  Reasons.js by Dave Kinkead
-//  Copyright (c) 2017-2018 University of Queensland
+//  Copyright 2017-2019 University of Queensland
 //  Available under the MIT license
 
 'use strict'
@@ -14,6 +14,7 @@ const fontSize     = 16
 const cornerRadius = 4
 const rgbFocused   = '81,36,122'
 const rgbDefault   = '0,0,0'
+
 
 let graph = {}
 
