+++ lib/utils.js
@@ -1,5 +1,5 @@
 //  Reasons.js by Dave Kinkead
-//  Copyright (c) 2017-2018 University of Queensland
+//  Copyright 2017-2019 University of Queensland
 //  Available under the MIT license
 
 module.exports = {
