+++ lib/graph.js
@@ -1,5 +1,5 @@
 //  Reasons.js by Dave Kinkead
-//  Copyright (c) 2017-2018 University of Queensland
+//  Copyright 2017-2019 University of Queensland
 //  Available under the MIT license
 
 'use strict'
@@ -52,8 +52,7 @@ Graph.prototype.add = function (element) {
     //  If A B & C both already support D
     //  and a new edge is added from A to B or vice versa
     //  then the relationships should be merged [A,B] -> D
-    //  and C -> D kept unchanged  
-    // console.log(this.children(element.from[0]))
+    //  and C -> D kept unchanged      
     let commonChildren = Utils.intersection(
       Utils.flatten(element.from.map(e => this.children(e))), 
       this.children(element.to)
