### 2019-03-11 bc55185
Prevent nodes being dropped on edges
