+++ lib/graph.js
@@ -48,6 +48,12 @@ Graph.prototype.add = function (element) {
     this.push(element)
   } else {
 
+    //  Sanity check to ensure that edges only join nodes
+    let to = this.find((el) => el.id == element.to)
+    if (to && to.isEdge()) {
+      return false
+    }
+
     //  Edges can connect independent or conjoined reasons. 
     //  If A B & C both already support D
     //  and a new edge is added from A to B or vice versa
