### 2019-03-11 8cf9546
Prevent duplicate edges from being added
