+++ reasons.js
@@ -260,7 +260,7 @@ Graph.prototype.add = function (element) {
 
     //  Sanity check to ensure that edges only join nodes
     let to = this.find((el) => el.id == element.to)
-    if (to && to.isEdge()) {
+    if (to && to.isEdge() || this.hasDuplicate(element)) {
       return false
     }
 
@@ -448,6 +448,27 @@ Graph.prototype.children = function (id) {
     return Utils.flatten([el.from]).map(el => el.id || el).indexOf(id) > -1
   }).map(el => el.to)
     .map(el => this.find(i => i == el || i.id == el)))
+}
+
+
+/**
+ * Determine if the proposed element is a dupliucate
+ *  Returns boolean
+ *
+ *  @params edge object
+ */
+Graph.prototype.hasDuplicate = function (el) {
+  Element.mixin(el)
+  let dupe = false
+
+  if (el.isEdge()) {
+    this.edges().forEach((edge) => {
+      if (el.to === edge.to && el.from.toString() === edge.from.toString()) {
+        dupe = true
+      }
+    })
+  }
+  return dupe
 }
 },{"./element":3,"./utils":8}],5:[function(require,module,exports){
 //  Reasons.js by Dave Kinkead
