### 2019-03-11 c90b34c
Fix backspace preventDefault issue
