+++ reasons.js
@@ -661,6 +661,15 @@ function addEventListeners (argumentMap) {
   })
 
 
+  //  Close modal if click occurs outside text box
+  window.addEventListener('click', (event) => {
+    
+    if (argumentMap.editMode && event.target.id === 'reason-overlay') {
+      removeOverlay(argumentMap)
+    }
+  })
+
+
   window.addEventListener('keydown', (event) => {
 
     if (argumentMap.editMode) {
@@ -715,11 +724,14 @@ function addEventListeners (argumentMap) {
       }
 
       //  Delete a selected element on `backspace` or `delete`
-      if (selected && (Keycode.isEventKey(event, 'delete') || Keycode.isEventKey(event, 'backspace'))) {
-        event.preventDefault()
-        argumentMap.graph.remove(selected)
-        argumentMap.dirty = true
-      }      
+      if (Keycode.isEventKey(event, 'delete') || Keycode.isEventKey(event, 'backspace')) {
+        if (!argumentMap.editMode) event.preventDefault()
+
+        if (selected) {
+          argumentMap.graph.remove(selected)
+          argumentMap.dirty = true          
+        }
+      }
     }
 
     redraw(argumentMap)
