### 2019-03-11 94af561
Fix bug - can create edge from edge to node
