+++ lib/view.js
@@ -15,7 +15,7 @@ const cornerRadius = 4
 const rgbFocused   = '81,36,122'
 const rgbDefault   = '0,0,0'
 
-
+let dpr = 1
 let graph = {}
 
 
@@ -31,6 +31,8 @@ module.exports = (function () {
    *  @params mapper  The argument map to provide a view for
    */
   function init (argument) {
+    dpr = window.devicePixelRatio || 1
+
     let domBB = argument.DOM.getBoundingClientRect()
     let canvas = Utils.buildNode(
       'canvas', 
@@ -39,7 +41,8 @@ module.exports = (function () {
     )
 
     argument.DOM.appendChild(canvas)
-    argument.context = canvas.getContext('2d')
+    argument.context = canvas.getContext('2d', {alpha: true})
+    // argument.context.scale(dpr ,dpr)
   }
 
 
@@ -85,10 +88,9 @@ module.exports = (function () {
   }
 
   function resize (argument) {
-    argument.DOM.children[1].width = argument.DOM.clientWidth - argument.DOM.clientLeft 
-    argument.DOM.children[1].height = argument.DOM.clientHeight - argument.DOM.clientTop 
+    argument.DOM.children[1].width = (argument.DOM.clientWidth - argument.DOM.clientLeft) 
+    argument.DOM.children[1].height = (argument.DOM.clientHeight - argument.DOM.clientTop) 
   }
-
 
   return {
     init,
@@ -96,8 +98,7 @@ module.exports = (function () {
     zero,
     resize
   }
-
-})();
+})()
 
 
 /**
@@ -120,7 +121,7 @@ function draw_node (node, context) {
   const opacity = (node.focused) ? 0.9 : (node.hovering) ? 0.75 : 0.5
 
   //  recalculate the height
-  node.height = text.length * fontSize + fontSize * 2.5
+  node.height = (text.length * fontSize + fontSize * 2.5) 
   resize(node)
 
   //  clear a white rectangle for background
