+++ lib/graph.js
@@ -49,8 +49,7 @@ Graph.prototype.add = function (element) {
   } else {
 
     //  Sanity check to ensure that edges only join nodes
-    let to = this.find((el) => el.id == element.to)
-    if (to && to.isEdge() || this.hasDuplicate(element)) {
+    if (this.hasDuplicate(element) || this.isFromEdge(element) || this.isToEdge(element)) {
       return false
     }
 
@@ -259,4 +258,35 @@ Graph.prototype.hasDuplicate = function (el) {
     })
   }
   return dupe
+}
+
+
+/**
+ * Determine if the proposed Edge is from an Edge
+ *  Returns boolean
+ *
+ *  @params edge object
+ */
+Graph.prototype.isFromEdge = function (element) {
+  var fromEdge = false
+  element.from.forEach((el) => {
+    var match = this.edges().find((e) => e.id == el)
+    if (match) fromEdge = true
+  })
+
+  return fromEdge
+}
+
+
+/**
+ * Determine if the proposed Edge is to an Edge
+ *  Returns boolean
+ *
+ *  @params edge object
+ */
+Graph.prototype.isToEdge = function (element) {
+  let to = this.find((el) => el.id == element.to)
+  if (to && to.isEdge()) {
+    return true
+  }
 }
