### 2019-03-11 06a0275
Fix bug - can create edge from edge to node
