### 2019-03-11 822cf8f
Move multi edge intersection closer to conclusion
