### 2019-03-13 d36c925
Update stroke width
