+++ lib/view.js
@@ -174,8 +174,8 @@ function draw_edge (edge, context) {
   context.stroke()
 
   //  text stroke
-  let textWidth = context.measureText(edge.type).width + 5
-  context.clearRect(edge.center.x-textWidth/2, edge.center.y-15, textWidth, 20)
+  let textWidth = context.measureText(edge.type).width + padding
+  context.clearRect(edge.center.x-textWidth/2, edge.center.y-15, textWidth, 25)
 
   //  label
   context.fillStyle = 'rgba('+rgb+',0.8)'
