### 2019-03-14 d55be55
Refactor metakey press event so control-z works on windows
