### 2019-03-14 davekinkead/reasons#2 @davekinkead
Sorry for the delay. I was unable to reproduce the issue on a VirtualBox and have only just had access to a windows machine. `ctrl z` and `ctrl y` now work and I've manually tested windows 10 + chrome & windows 10 + firefox.
