### 2019-03-14 #1044 @davekinkead
All fixed now & good to go. Sorry for the delay.
