### 2019-05-12 davekinkead/reasons#5 @yochannah
Example code on readme produces a 0-height canvas element
1. Clone and NPM install the package
2. Copy/paste the demo HTML and serve locally
3. This results in a 0px high canvas element and no errors on the console:

The graph elements *are* present on the screen if I edit the DOM to give the canvas element height.

Tested on both ff and chrome, on 1.0.0 and the master branch.
