### 2019-05-12 #1044 @yochannah
Okay, I've run through the review steps myself. @davekinkead I'll be happy to accept this, one this issue is fixed with the demo code in the repo:

https://github.com/davekinkead/reasons/issues/5

I'll also give the paper a quick once-over again now.
