### 2019-05-12 4c4e341
Add default height to canvas. Closes #5
