### 2019-05-12 1a86046
Add DOIs to references
