+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={Academically adrift: Limited learning on college campuses},
   author={Arum, Richard and Roksa, Josipa},
   year={2011},
-  publisher={University of Chicago Press}
+  publisher={University of Chicago Press},
+  doi={https://doi.org/10.7208/chicago/9780226028576.001.0001}
 }
 
 @article{butchart2009improving,
@@ -11,7 +12,8 @@
   journal={Australasian Journal of Educational Technology},
   volume={25},
   number={2},
-  year={2009}
+  year={2009},
+  doi={https://doi.org/10.14742/ajet.1154}
 }
 
 @article{davies2009computer,
@@ -22,7 +24,9 @@
   number={6},
   pages={799},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/s10734-009-9226-9}
+
 }
 
 @article{dwyer2012evaluation,
@@ -33,7 +37,8 @@
   number={3},
   pages={219--244},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/s11409-012-9092-1}
 }
 
 @article{govier1992good,
@@ -44,7 +49,8 @@
   number={4},
   pages={393--409},
   year={1992},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1111/j.1467-9973.1992.tb00551.x}
 }
 
 @article{harrell2004improvement,
@@ -53,7 +59,8 @@
   journal={Carnegie Mellon University. Retrieved May},
   volume={9},
   pages={2007},
-  year={2004}
+  year={2004},
+  url={https://www.cmu.edu/dietrich/philosophy/docs/harrell/Improving_Critical_Thinking_Skills.pdf}
 }
 
 @article{mulnix2012thinking,
@@ -64,13 +71,15 @@
   number={5},
   pages={464--479},
   year={2012},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={https://doi.org/10.1111/j.1469-5812.2010.00673.x}
 }
 
 @book{toulmin2003uses,
   title={The uses of argument},
   author={Toulmin, Stephen E},
   year={2003},
+  doi = {10.1017/cbo9780511840005.005},
   publisher={Cambridge university press}
 }
 
@@ -81,6 +90,7 @@
   volume={27},
   number={2},
   pages={95--116},
+  doi = {10.5840/teachphil200427213},
   year={2004}
 }
 
@@ -91,6 +101,7 @@
   volume={2},
   number={1},
   pages={85--90},
+  url={https://sites.google.com/site/timvangelder/publications-1/argument-mapping-with-reason-able/ArgumentMappingwithReasonAble-APANewsletter.pdf},
   year={2002}
 }
 
@@ -99,5 +110,6 @@
   author={Wigmore, John Henry},
   volume={1},
   year={1913},
-  publisher={Little, Brown,}
+  publisher={Little, Brown,},
+  doi={https://doi.org/10.2307/1275020}
 }
