### 2019-05-12 #1044 @davekinkead
Hi @yochannah - https://github.com/davekinkead/reasons/issues/5 has been actioned & DOIs added

Thanks
