### 2019-05-12 #1044 @yochannah
Hmm, so whedon can see that the references are there since it passes the check references step, but it's not generating links in the paper itself for some reason. @openjournals/joss-eics do we need to do anything to fix this?
