### 2019-05-12 #1044 @yochannah
@davekinkead Thanks, the bug is fixed and looking good!

One small extra thing I've noticed in the references:

> Toulmin, S. E. (1958). The uses of argument. Cambridge university press.

I suspect that Cambridge University Press should be capitalised. It's probably worth giving the manuscript a final proofread if you feel you need to, as well. 👍

Once the references thing above is all ironed out I'll ask you to make an archive on Zenodo or similar, but let's wait until we're sure we don't need to make any changes first.
