### 2019-05-12 #1044 @labarba
@davekinkead — Try putting _just_ the DOI, i.e., the alphanumeric identifier, not the full URL, into the `doi` field of the `.bib` file.
