+++ paper/paper.md
@@ -29,13 +29,13 @@ bibliography: paper.bib
 
 There is growing recognition globally of the need to teach Critical Thinking as part of formal schooling and of its importance to the “knowledge economies” of the future.  Yet international research demonstrates that without explicit instruction in critical thinking, undergraduate education often results in little to no gains in critical thinking, analytic reasoning, and other "higher level" skills (Harrell 2004, Arum & Roksa 2011). 
 
-One very effective way to improving critical thinking is through argument mapping — the visual representation of an argument’s logical structure.  Argument mapping in paper form is common in philosophy courses and has a pedagogical pedigree that can be traced back to Wigmore (1913), Toulmin (1958), and Govier (1988).  Argument can improve critical thinking skills by offering students an opportunity to engage in _metacogntive evaluation_ -  evaluating the quality of their own, and other's, reasoning.
+One very effective way to improving critical thinking is through argument mapping — the visual representation of an argument’s logical structure.  Argument mapping in paper form is common in philosophy courses and has a pedagogical pedigree that can be traced back to Wigmore (1913), Toulmin (1958), and Govier (1988).  Argument mapping can improve critical thinking skills by offering students an opportunity to engage in _metacogntive evaluation_ -  evaluating the quality of their own, and other's, reasoning.
 
-Digital argument mapping as an educational tool has been validated by van Gelder (2002), Butchart et al (2009), and Mulnix (2012). Dwyer, Hogan, & Stewart (2012) demonstrated that argument mapping improves concept recall compared with textual analysis; Twardy (2004: 2), that it produces cognitive gains three times that of other methods; and van Gelder (2005: 45) that the cognitive gains from one semester of explicit argument mapping are equivalent to that of an entire undergraduate degree.  
+Digital argument mapping as an educational tool has been validated by van Gelder (2002), Butchart et al (2009), and Mulnix (2012). Dwyer, Hogan, & Stewart (2012) demonstrated that argument mapping improves concept recall compared with textual analysis; Twardy (2004: 2) that it produces cognitive gains three times that of other methods; and van Gelder (2005: 45) that the cognitive gains from one semester of explicit argument mapping are equivalent to that of an entire undergraduate degree.  
 
 Unfortunately, argument mapping is rarely used outside of philosophy classes owing either to a lack of instructor expertise or availability of tools appropriate to non-philosophical pedagogies.  Current digital argument mapping tools are either desktop software, limiting their ability to be integrated into online courseware, or propriety and tighly coupled, limiting their access and extensibility.
 
-`Reasons` seeks to bridge this gap by offering an open-source, loosely-coupled, web-based argument mapping library that can be integrated into a range of online coursewares and websites.  The javascript library can be embedded into any HTML page and allows users to create, edit, share, and export argument maps (see http://reasons.io for an example).  The API is designed to permit the integration of the three stages of informal logical analysis -- identification of truth claims within arguments, the analysis of logical structure, and synthesis of logcial structure into writen form.
+`Reasons` seeks to bridge this gap by offering an open-source, loosely-coupled, web-based argument mapping library that can be integrated into a range of online coursewares and websites.  The javascript library can be embedded into any HTML page and allows users to create, edit, share, and export argument maps (see https://reasons.io for an example).  The API is designed to permit the integration of the three stages of informal logical analysis -- identification of truth claims within arguments, the analysis of logical structure, and synthesis of logcial structure into writen form.
 
[…]
