+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Arum, Richard and Roksa, Josipa},
   year={2011},
   publisher={University of Chicago Press},
-  doi={https://doi.org/10.7208/chicago/9780226028576.001.0001}
+  doi={10.7208/chicago/9780226028576.001.0001}
 }
 
 @article{butchart2009improving,
@@ -13,7 +13,7 @@
   volume={25},
   number={2},
   year={2009},
-  doi={https://doi.org/10.14742/ajet.1154}
+  doi={10.14742/ajet.1154}
 }
 
 @article{davies2009computer,
@@ -25,7 +25,7 @@
   pages={799},
   year={2009},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s10734-009-9226-9}
+  doi={10.1007/s10734-009-9226-9}
 
 }
 
@@ -38,7 +38,7 @@
   pages={219--244},
   year={2012},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s11409-012-9092-1}
+  doi={10.1007/s11409-012-9092-1}
 }
 
 @article{govier1992good,
@@ -50,7 +50,7 @@
   pages={393--409},
   year={1992},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1111/j.1467-9973.1992.tb00551.x}
+  doi={10.1111/j.1467-9973.1992.tb00551.x}
 }
 
 @article{harrell2004improvement,
@@ -72,15 +72,17 @@
   pages={464--479},
   year={2012},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1111/j.1469-5812.2010.00673.x}
+  doi={10.1111/j.1469-5812.2010.00673.x}
 }
 
-@book{toulmin2003uses,
-  title={The uses of argument},
-  author={Toulmin, Stephen E},
-  year={2003},
+@incollection{toulmin2003uses,
   doi = {10.1017/cbo9780511840005.005},
-  publisher={Cambridge university press}
+  url = {https://doi.org/10.1017%2Fcbo9780511840005.005},
+  publisher = {Cambridge University Press},
+  pages = {11--40},
+  author = {Stephen E. Toulmin},
+  title = {Fields of Argument and Modals},
+  booktitle = {The Uses of Argument}
 }
 
 @article{twardy2004argument,
@@ -111,5 +113,5 @@
   volume={1},
   year={1913},
   publisher={Little, Brown,},
-  doi={https://doi.org/10.2307/1275020}
+  doi={10.2307/1275020}
 }
