+++ paper/paper.md
@@ -27,11 +27,11 @@ bibliography: paper.bib
 
 # Summary
 
-There is growing recognition globally of the need to teach Critical Thinking as part of formal schooling and of its importance to the “knowledge economies” of the future.  Yet international research demonstrates that without explicit instruction in critical thinking, undergraduate education often results in little to no gains in critical thinking, analytic reasoning, and other "higher level" skills (Harrell 2004, Arum & Roksa 2011). 
+There is growing recognition globally of the need to teach Critical Thinking as part of formal schooling and of its importance to the “knowledge economies” of the future.  Yet international research demonstrates that without explicit instruction in critical thinking, undergraduate education often results in little to no gains in critical thinking, analytic reasoning, and other "higher level" skills [@harrell2004improvement, @arum2011academically]. 
 
-One very effective way to improving critical thinking is through argument mapping — the visual representation of an argument’s logical structure.  Argument mapping in paper form is common in philosophy courses and has a pedagogical pedigree that can be traced back to Wigmore (1913), Toulmin (1958), and Govier (1988).  Argument mapping can improve critical thinking skills by offering students an opportunity to engage in _metacogntive evaluation_ -  evaluating the quality of their own, and other's, reasoning.
+One very effective way to improving critical thinking is through argument mapping — the visual representation of an argument’s logical structure.  Argument mapping in paper form is common in philosophy courses and has a pedagogical pedigree that can be traced back to Wigmore [@wigmore1913principles], Toulmin [@toulmin2003uses], and Govier [@govier1992good].  Argument mapping can improve critical thinking skills by offering students an opportunity to engage in _metacogntive evaluation_ -  evaluating the quality of their own, and other's, reasoning.
 
-Digital argument mapping as an educational tool has been validated by van Gelder (2002), Butchart et al (2009), and Mulnix (2012). Dwyer, Hogan, & Stewart (2012) demonstrated that argument mapping improves concept recall compared with textual analysis; Twardy (2004: 2) that it produces cognitive gains three times that of other methods; and van Gelder (2005: 45) that the cognitive gains from one semester of explicit argument mapping are equivalent to that of an entire undergraduate degree.  
+Digital argument mapping as an educational tool has been validated by van Gelder [@van2002argument], Butchart et al [@butchart2009improving], and Mulnix [@mulnix2012thinking]. Dwyer, Hogan, & Stewart [dwyer2012evaluation] demonstrated that argument mapping improves concept recall compared with textual analysis; Twardy [twardy2004argument: p2] that it produces cognitive gains three times that of other methods; and van Gelder [@vangelder_2005 :45] that the cognitive gains from one semester of explicit argument mapping are equivalent to that of an entire undergraduate degree.  
 
 Unfortunately, argument mapping is rarely used outside of philosophy classes owing either to a lack of instructor expertise or availability of tools appropriate to non-philosophical pedagogies.  Current digital argument mapping tools are either desktop software, limiting their ability to be integrated into online courseware, or propriety and tighly coupled, limiting their access and extensibility.
 
