+++ paper/paper.bib
@@ -107,6 +107,19 @@
   year={2002}
 }
 
+@article{vangelder_2005,
+  doi = {10.3200/ctch.53.1.41-48},
+  url = {https://doi.org/10.3200%2Fctch.53.1.41-48},
+  year = 2005,
+  month = {jan},
+  publisher = {Informa {UK} Limited},
+  volume = {53},
+  number = {1},
+  pages = {41--48},
+  author = {Tim van Gelder},
+  title = {Teaching Critical Thinking: Some Lessons From Cognitive Science},
+  journal = {College Teaching}
+}
 @book{wigmore1913principles,
   title={The principles of judicial proof: as given by logic, psychology, and general experience, and illustrated in judicial trials},
   author={Wigmore, John Henry},
