### 2019-05-12 #1044 @labarba
Just before the references, you have a link to the code repository. We include a link to the repo on the margin decorators (first page), so we ask that you don't add an additional link in the text or references. Can you edit that?
