+++ paper/paper.bib
@@ -29,7 +29,7 @@
 
 }
 
-@article{dwyer2012evaluation,
+@article{dwyer2012,
   title={An evaluation of argument mapping as a method of enhancing critical thinking performance in e-learning environments},
   author={Dwyer, Christopher P and Hogan, Michael J and Stewart, Ian},
   journal={Metacognition and Learning},
@@ -80,6 +80,7 @@
   url = {https://doi.org/10.1017%2Fcbo9780511840005.005},
   publisher = {Cambridge University Press},
   pages = {11--40},
+  year = {1958},
   author = {Stephen E. Toulmin},
   title = {Fields of Argument and Modals},
   booktitle = {The Uses of Argument}
