### 2019-05-12 #1044 @labarba
For your last reference, the DOI points to the book review. Did you mean this?
