+++ paper/paper.bib
@@ -117,7 +117,7 @@
   volume = {53},
   number = {1},
   pages = {41--48},
-  author = {Tim van Gelder},
+  author={Van Gelder, Tim},
   title = {Teaching Critical Thinking: Some Lessons From Cognitive Science},
   journal = {College Teaching}
 }
