### 2019-05-12 #1044 @labarba
Regarding Wigmore (1913): you are citing the book, not the review of the book, so that DOI is not correct. However, I did find that the book is now in the public domain and you can find an online archive at https://archive.org/details/principlesofjudi00wigm/ — maybe you could add that URL.
