+++ paper/paper.bib
@@ -127,5 +127,5 @@
   volume={1},
   year={1913},
   publisher={Little, Brown,},
-  doi={10.2307/1275020}
+  url={https://archive.org/details/principlesofjudi00wigm/}
 }
