### 2019-05-12 #1044 @davekinkead
Comments

https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491593171
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491593290
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491593585
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491595545
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491595939
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491596274
https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491597090

have now been actioned. Thanks for your patience @yochannah @labarba
