### 2019-05-12 #1044 @labarba
Oops. Please now make a tagged release, and report the version number here, then make a deposit of your full repository on an archive service like Zenodo, and report the DOI here.
