### 2019-05-12 #1044 @davekinkead
Hi @labarba

Tagged release https://github.com/davekinkead/reasons/releases/tag/v1.0.1
Zendo https://zenodo.org/record/2766003
