### 2019-05-12 #1044 @labarba
I have now made this edit to the top post: https://github.com/openjournals/joss-reviews/issues/1044#issuecomment-491600234
