### 2019-05-12 #1044 @labarba
Congratulations, @davekinkead, your JOSS paper is now published!

Many thanks to @yochannah, @andytwoods for their work on this submission 🙏
